requirements.txt
pandas==2.0.3
numpy==1.24.3
pandas-ta>=0.3.14b0
pyupbit>=0.2.33
python-dotenv>=1.0.0
loguru==0.7.2
rich==13.7.0
httpx>=0.24.1
pytest==7.4.3
python-telegram-bot>=20.4
cryptography>=41.0.3
yfinance==0.2.36
pytest-cov==4.1.0 